review: widen resolved_outer_length px; pin checked in py binding; CRLF bat
- resolved_outer_length still took int16 px, re-narrowing the widened
  ComputedStyle width/height for absolutely/fixed-positioned boxes.
- virtual_string_list keep_alive pinned the item_size double (no-op)
  instead of the checked IndexSelection; now pins args 5 and 6.
- rust_check.bat committed as CRLF + .gitattributes rule (*.bat eol=crlf).

Addresses CodeRabbit review on #48.

+++ src/renderer/dom/document_restyle.cpp
@@ -347,7 +347,9 @@ int resolve_position_edge(std::int16_t value, bool is_pct, int basis) {
                     (static_cast<double>(value) / 10000.0)));
 }
 
-int resolved_outer_length(std::int16_t px,
+// px is int32: ComputedStyle sizing fields widened past int16 (virtual-list
+// spacers reach millions of px); narrowing here would re-wrap them.
+int resolved_outer_length(std::int32_t px,
                           std::int16_t pct_x100,
                           int basis,
                           int fallback,
